checking for ambiguity and removing dead code
git-svn-id: https://phycas.svn.sourceforge.net/svnroot/phycas/trunk@1107 b774380d-c90d-0410-997c-aec342e5d6fd

+++ phycas/src/tip_data.cpp
@@ -85,7 +85,7 @@ TipData::TipData(
         for (unsigned i = 0; i < nPatterns; ++i)
         	{
         	const int8_t sc = state_codes[i];
-        	if (sc < 0 || sc >= nStates)
+        	if (sc < 0 || (int)sc >= (int)nStates)
         		throw XLikelihood("Sorry, we currently do not support data sets with ambiguity or gaps when you are using uniformization-based methods");
         	}
         }
